kafka-payment-labs/requirements.txt
# Kafka Payment Processing Labs - Python Dependencies

# Kafka client
confluent-kafka[avro,json,protobuf]>=2.3.0

# Schema Registry and Avro
fastavro>=1.9.0
jsonschema>=4.20.0

# Web framework for REST API
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0  # For testing APIs

# Database connectivity
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
click>=8.1.0  # For CLI tools

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Development tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Data generation for testing
faker>=20.1.0